+++ src/spyglass/common/populate_all_common.py
@@ -42,7 +42,7 @@ from spyglass.common.common_usage import InsertError
 from spyglass.settings import base_dir
 from spyglass.utils import logger
 from spyglass.utils.dj_helper_fn import declare_all_merge_tables
-from spyglass.utils.dj_mixin import SpyglassIngestion
+from spyglass.utils.dj_mixin import IngestionMixin
 
 
 def log_insert_error(
@@ -107,7 +107,7 @@ def single_transaction_make(
             config_name = _get_config_name(table())
             table_config = config.get(config_name, dict())
 
-            if isinstance(table(), SpyglassIngestion):
+            if isinstance(table(), IngestionMixin):
                 try:
                     table().insert_from_nwbfile(
                         nwb_file_name, config=table_config