+++ src/spyglass/utils/__init__.py
@@ -1,15 +1,15 @@
 from spyglass.utils.dj_merge_tables import _Merge
 from spyglass.utils.dj_mixin import (
-    SpyglassIngestion,
     SpyglassAnalysis,
     SpyglassMixin,
     SpyglassMixinPart,
 )
 from spyglass.utils.logging import logger
+from spyglass.utils.mixins.ingestion import IngestionMixin
 
 __all__ = [
     "SpyglassAnalysis",
-    "SpyglassIngestion",
+    "IngestionMixin",
     "SpyglassMixin",
     "SpyglassMixinPart",
     "_Merge",